Make progress with parsing config and enums. TEMP COMMIT.
